requirements.txt
fastapi
uvicorn
psutil
requests
python-dotenv
appdirs
discord.py
